+++ src/components/SupervisorBargeCoachButton/SupervisorBargeCoachButton.Component.js
@@ -1,40 +1,11 @@
 import * as React from 'react';
-import { IconButton, TaskHelper, withTheme } from '@twilio/flex-ui';
-import styled from 'react-emotion';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
+import { IconButton, TaskHelper } from '@twilio/flex-ui';
 
-import { syncClient, conferenceClient, localCacheClient } from '../services';
-import { Actions as BargeCoachStatusAction } from '../states/BargeCoachState';
-import { logger } from '../utils';
+import { syncClient, conferenceClient } from '../../services';
+import { logger } from '../../utils';
+import { ButtonContainer, buttonStyle, buttonStyleActive } from './SupervisorBargeCoachButton.Style';
 
-const ButtonContainer = styled('div')`
-  display: flex;
-  justify-content: center;
-  margin-bottom: 6px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  align-items: center;
-  text-align: center;
-`;
-
-const buttonStyleActive = {
-  width: '44px',
-  height: '44px',
-  marginLeft: '6px',
-  marginRight: '6px',
-  color: 'limegreen',
-};
-
-const buttonStyle = {
-  width: '44px',
-  height: '44px',
-  marginLeft: '6px',
-  marginRight: '6px',
-};
-
-class SupervisorBargeCoachButton extends React.Component {
+export default class SupervisorBargeCoachButton extends React.Component {
   /*
    * On click we will be pulling the conference SID and supervisorSID
    * to trigger Mute / Unmute respectively for that user - muted comes from the redux store
@@ -244,60 +215,3 @@ class SupervisorBargeCoachButton extends React.Component {
     );
   }
 }
-
-/*
- * Getting the Supervisor's workerSID so we can use it later, the Agent's workerSID (stickyWorker) we are monitoring
- * This is specific to coaching to ensure we are unmuting the correct worker, if there are multiple agents on the call
- */
-const mapStateToProps = (state) => {
-  const myWorkerSID = state?.flex?.worker?.worker?.sid;
-  const agentWorkerSID = state?.flex?.supervisor?.stickyWorker?.worker?.sid;
-  const supervisorFN = state?.flex?.worker?.attributes?.full_name;
-  logger.log(`sticky worker = ${agentWorkerSID}`);
-
-  /*
-   * Also pulling back the states from the redux store as we will use those later
-   * to manipulate the buttons
-   */
-  const customReduxStore = state?.['barge-coach'].bargecoach;
-  const { muted } = customReduxStore;
-  const { barge } = customReduxStore;
-  const { enableBargeinButton } = customReduxStore;
-  const { coaching } = customReduxStore;
-  const { enableCoachButton } = customReduxStore;
-  const { coachingStatusPanel } = customReduxStore;
-
-  const teamViewPath = state?.flex?.router?.location?.pathname;
-
-  /*
-   * Storing teamViewPath and agentSyncDoc to browser cache to help if a refresh happens
-   * will use this in the main plugin file to invoke an action to reset the monitor panel
-   * and clear the Agent's Sync Doc
-   */
-  if (teamViewPath !== null) {
-    localCacheClient.setTeamViewPath(teamViewPath);
-    localCacheClient.setAgentSyncDoc(`syncDoc.${agentWorkerSID}`);
-  }
-
-  return {
-    myWorkerSID,
-    agentWorkerSID,
-    supervisorFN,
-    muted,
-    barge,
-    enableBargeinButton,
-    coaching,
-    enableCoachButton,
-    coachingStatusPanel,
-  };
-};
-
-/*
- * Mapping dispatch to props as I will leverage the setBargeCoachStatus
- * to change the properties on the redux store, referenced above with this.props.setBargeCoachStatus
- */
-const mapDispatchToProps = (dispatch) => ({
-  setBargeCoachStatus: bindActionCreators(BargeCoachStatusAction.setBargeCoachStatus, dispatch),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(withTheme(SupervisorBargeCoachButton));
