+++ src/components/index.js
@@ -1,1 +1,2 @@
 export { default as SupervisorPrivateModeButton } from './SupervisorPrivateModeButton';
+export { default as SupervisorBargeCoachButton } from './SupervisorBargeCoachButton';
